EOS-16521 : Convert sspl_post_install shell script to python
